pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "word-to-markdown"
version = "0.1.0"
description = "Convert Word .doc/.docx files into structured Markdown."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-docx>=1.2.0",
]

[project.scripts]
word-to-markdown = "word_to_markdown.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

